Make integration tests recursive (#3)
This probably constitutes a breaking change. Currently services using
this Makefile will not run all of the integration tests because the
recursive flag is missing. I should have noticed this during testing.

It's safe to always add the flag for our services as they all have a
non-flat directory structure for the tests.